+++ demo/lib/demo_web/router.ex
@@ -22,6 +22,7 @@ defmodule DemoWeb.Router do
       live "/tables", Live.TablesLive, :index
       live "/tables-sizing", Live.TablesSizingLive, :index
       live "/tables-responsive", Live.TablesResponsiveLive, :index
+      live "/table-filters", Live.TableFiltersLive, :index
       live "/modals", Live.ModalsLive, :index
       live "/modal-dialogs", Live.ModalDialogsLive, :index
       live "/popconfirm", Live.PopconfirmLive, :index